fixing cache

+++ src/Tests/EntityStorageTest.php
@@ -321,32 +321,31 @@ class EntityStorageTest extends MultiversionWebTestBase {
 
       // Test workspace.
 
-      if ($entity_type_id == 'user') {
-        $info['info']['name'] = $this->randomMachineName();
-      }
-      $entity = $storage->create($info['info']);
-      $entity->save();
-      $entity_id = $entity->id();
-      $this->assertEqual($entity->workspace->target_id, 1, "The workspace reference was saved for $entity_type_id.");
-      $record = db_select($info['data_table'], 'e')
-        ->fields('e')
-        ->condition('e.' . $info['id'], $entity->id())
-        ->condition('e.' . $info['revision_id'], $entity->getRevisionId())
-        ->execute()
-        ->fetchObject();
-      $this->assertEqual($record->workspace, 1, "The workspace reference was stored for saved $entity_type_id.");
+      if ($entity_type->get('workspace') !== FALSE) {
+        $entity = $storage->create($info['info']);
+        $entity->save();
+        $entity_id = $entity->id();
+        $this->assertEqual($entity->workspace->target_id, 1, "The workspace reference was saved for $entity_type_id.");
+        $record = db_select($info['data_table'], 'e')
+          ->fields('e')
+          ->condition('e.' . $info['id'], $entity->id())
+          ->condition('e.' . $info['revision_id'], $entity->getRevisionId())
+          ->execute()
+          ->fetchObject();
+        $this->assertEqual($record->workspace, 1, "The workspace reference was stored for saved $entity_type_id.");
 
-      $entities = $storage->loadMultiple([$entity_id]);
-      $storage->delete($entities);
-      $entity = $storage->loadDeleted($entity_id);
-      $this->assertEqual($entity->workspace->target_id, 1, "The workspace reference is retained for deleted $entity_type_id.");
-      $record = db_select($info['data_table'], 'e')
-        ->fields('e')
-        ->condition('e.' . $info['id'], $entity->id())
-        ->condition('e.' . $info['revision_id'], $entity->getRevisionId())
-        ->execute()
-        ->fetchObject();
-      $this->assertEqual($record->workspace, 1, "The workspace reference was stored for deleted $entity_type_id.");
+        $entities = $storage->loadMultiple([$entity_id]);
+        $storage->delete($entities);
+        $entity = $storage->loadDeleted($entity_id);
+        $this->assertEqual($entity->workspace->target_id, 1, "The workspace reference is retained for deleted $entity_type_id.");
+        $record = db_select($info['data_table'], 'e')
+          ->fields('e')
+          ->condition('e.' . $info['id'], $entity->id())
+          ->condition('e.' . $info['revision_id'], $entity->getRevisionId())
+          ->execute()
+          ->fetchObject();
+        $this->assertEqual($record->workspace, 1, "The workspace reference was stored for deleted $entity_type_id.");
+      }
     }
 
     // Test saving entities in a different workspace.
@@ -361,10 +360,10 @@ class EntityStorageTest extends MultiversionWebTestBase {
     $this->workspaceManager->setActiveWorkspace($workspace);
 
     foreach ($this->entityTypes as $entity_type_id => $info) {
-      if ($info->get('workspace') !== FALSE) {
         $storage = $this->entityManager->getStorage($entity_type_id);
         $entity = $storage->create($info['info']);
         $entity->save();
+      if ($entity->getEntityType()->get('workspace') !== FALSE) {
         $this->assertEqual($entity->workspace->target_id, $workspace->id(), "$entity_type_id was saved in new workspace.");
       }
     }
@@ -440,14 +439,17 @@ class EntityStorageTest extends MultiversionWebTestBase {
 
     // Save the same entities on target as new entities, but with the same UUID.
     $this->workspaceManager->setActiveWorkspace($target);
-    foreach ($this->entityTypes as $entity_type_id => $entityType) {
+    foreach ($this->entityTypes as $entity_type_id => $info) {
       /** @var \Drupal\Core\Entity\ContentEntityInterface $source_entity */
       foreach ($entities[$entity_type_id] as $source_entity) {
-        if ($entityType->get('workspace') !== FALSE) {
+        if ($source_entity->getEntityType()->get('workspace') !== FALSE) {
           $target_entity = clone $source_entity;
 
           // Reset the ID and assign the new workspace.
-          $target_entity->{$entityType['id']}->value = NULL;
+          $target_entity->{$info['id']}->value = NULL;
+
+          $target_entity->enforceIsNew(TRUE);
+          $target_entity->workspace->target_id = $target->id();
 
           // Save the new entity
           $target_entity->save();
